separate new coverage collection from normal running of selected tests
