+++ setup.py
@@ -18,6 +18,7 @@ setup(
             "tests_selector=tests_selector.select:main",
             "tests_selector_eval=tests_selector.evaluation.start:main",
             "tests_selector_run=tests_selector.run:main",
+            "tests_selector_run_and_update=tests_selector.run_and_update:main",
             "tests_selector_specific_without_remap=tests_selector.evaluation.specific_without_remap:main",
             "tests_selector_all_without_remap=tests_selector.evaluation.all_without_remap:main",
             "tests_selector_collect=tests_selector.collect:main",
